topic card design changes
