+++ app/assets/stylesheets/application.scss.erb
@@ -3150,8 +3150,15 @@ script.data-gratipay-username {
 }
 
 .topicFollow {
-    height: 48px;
-    line-height: 48px;
+    height: 24px;
+    line-height: 24px;
     cursor: pointer;
-    font-family: din-regular;
+    font-family: helvetica;
+    float: left;
+    width: 72px;
+    text-align: right;
+    padding: 12px 0;
+    color: #4fb5c0;
+    font-size: 13px;
+    font-weight: bold;
 }
