+++ frontend/src/components/TopicCard/Desc.js
@@ -25,7 +25,7 @@ class MdTextArea extends RIETextArea {
 class Desc extends Component {
   render = () => {
     const descHTML = (!this.props.desc && this.props.authorizedToEdit)
-      ? '<p>Click to add description...</p>'
+      ? '<p class="emptyDesc">Edit the description...</p>'
       : Util.mdToHTML(this.props.desc)
 
     if (this.props.authorizedToEdit) {
