+++ frontend/src/components/TopicCard/Follow.js
@@ -2,16 +2,17 @@ import React, { PropTypes, Component } from 'react'
 
 class Follow extends Component {
   render = () => {
-    const { isFollowing, onTopicFollow } = this.props
-    return <div className='topicFollow' onClick={onTopicFollow}>
-      {isFollowing ? 'Unfollow' : 'Follow'}
+    const { ActiveMapper, isFollowing, onTopicFollow } = this.props
+    return <div className='topicFollow' onClick={() => ActiveMapper && onTopicFollow()}>
+      {ActiveMapper ? isFollowing ? 'Unfollow' : 'Follow' : ''}
     </div>
   }
 }
 
 Follow.propTypes = {
   isFollowing: PropTypes.bool,
-  onTopicFollow: PropTypes.func
+  onTopicFollow: PropTypes.func,
+  ActiveMapper: PropTypes.object
 }
 
 export default Follow
